Test Fix
git-svn-id: http://svn.kolmisoft.com/mor/gui/branches/12@28807 19941b1f-7a74-413c-ab52-0de3e92e9a87

+++ app/helpers/application_helper.rb
@@ -70,7 +70,7 @@ module ApplicationHelper
     else
       nice_number_digits = (session and session[:nice_number_digits]) or Confline.get_value("Nice_Number_Digits")
       nice_number_digits = 2 if nice_number_digits == ''
-      n = sprintf("%0.#{nice_number_digits}f", number.to_d) if number
+      n = sprintf("%0.#{nice_number_digits}f", number.to_f) if number
       if session and session[:change_decimal]
         n = n.gsub('.', session[:global_decimal])
       end
